ocr-test/pyproject.toml
[project]
name = "ocr-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-cognitiveservices-vision-computervision>=0.9.1",
    "azure-core>=1.34.0",
    "boto3>=1.38.25",
    "google-cloud-vision>=3.10.1",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
]
